# swagger/swagger.yaml
swagger: "2.0"
info:
  description: This is a server for "fakelock" social network

  version: 0.0.1
  title: Swagger fakelock api
  contact:
    name: Lior Ezra
    email: admin@fakelock.com
host: localhost:3000
basePath: /api
...
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  BasicAuth:
    type: basic
paths:
...
  "/posts/{postId}":
    get:
      tags:
        - post
      summary: "get fakelock post by id"
      description: ""
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: path
          name: postId
          type: string
          required: true
          description: string ID of the post
      security:
        - BasicAuth: []
      responses:
        200:
          description: "return requested post by id"
        401:
          description: "TokenExpiredError/NoToken/TokenNotValid/JsonWebTokenError"
        404:
          description: "post not found"
        500:
          description: "Something broke!"
  "/posts":
    post:
      tags:
        - post
      summary: "add new post to fakelock"
      description: ""
      operationId: "createPost"
      consumes:
        - "multipart/form-data"
      produces:
        - "application/json"
      parameters:
        - in: "formData"
          name: "file"
          type: "file"
          required: true
        - in: "formData"
          name: "post"
          type: "string"
          required: true
          description: "should be json object(post model)"
      security:
        - BasicAuth: []
      responses:
        200:
          description: "return the id of the new post"
        400:
          description: "NotImage/some property missing"
        401:
          description: "TokenExpiredError/NoToken/TokenNotValid/JsonWebTokenError"
        406:
          description: "ImageSizeNotValid"
        500:
          description: "Something broke!"
    get:
      tags:
        - post
      summary: "get fakelock posts by filter query"
      description: ""
      operationId: "getPosts"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "filter"
          description: "filter object(model) to get posts after filter"
          required: true
          type: "string"
      security:
        - BasicAuth: []
      responses:
        200:
          description: "return posts by the filter query,can be empty"
        401:
          description: "TokenExpiredError/NoToken/TokenNotValid/JsonWebTokenError"
        500:
          description: "Something broke!"
...